requirements_test.txt
-r requirements_dev.txt
pytest-homeassistant-custom-component==0.13.263
ventaxiaiot>=0.1.2
propcache
